+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## [0.5.0]
 
 ### Added
-- True color to 256 color palette fallback by using gchalk ansistyles package
+- True color to 256 color palette fallback
 - Changed syntax of light colors (lightred ==> lred)
+- Added escape system [<content>]
 
 
 ## [0.4.0] - 2026-04-06